+++ include/cask/Task.hpp
@@ -108,7 +108,23 @@ public:
      */
     constexpr static Task<T,E> forPromise(const PromiseRef<T,E>& promise) noexcept;
 
-    static Task<T,E> forCancelableAndPromise(const CancelableRef& cancelable, const PromiseRef<T,E>& promise) noexcept;
+    /**
+     * A convenience method for a common task performed by observables.
+     * 
+     * This method takes a given cancelable and promise and returns a deferred that only
+     * completes after BOTH the given cancelable has completed AND the given promise has a
+     * success or error value. This ensures that callers cannot observe the result value
+     * until all relevant effects (including guaranteed shutdown effects and the like) have
+     * been evaluated. This is necessary, for example, when observing the result of an
+     * observable - the subscription must be properly completed AND a value of some kind
+     * must be obtained (via a promise) from an observer in the chain.
+     * 
+     * @param cancelable The cancelable background task to ensure is ended (e.g. an observable subscription)
+     * @param promise The promise which will provide the result value.
+     * @param sched The scheduler to use when composing these asynchronous operations.
+     * @return A deferred which is safeuly completed only when all relevant effects are completed.
+     */
+    static DeferredRef<T,E> runCancelableThenPromise(const CancelableRef& cancelable, const PromiseRef<T,E>& promise, const SchedulerRef& sched) noexcept;
 
     /**
      * Creates a task that will never finish evaluation.
@@ -451,24 +467,25 @@ constexpr Task<T,E> Task<T,E>::forPromise(const PromiseRef<T,E>& promise) noexce
 }
 
 template <class T, class E>
-Task<T,E> Task<T,E>::forCancelableAndPromise(const CancelableRef& cancelable, const PromiseRef<T,E>& promise) noexcept {
+DeferredRef<T,E> Task<T,E>::runCancelableThenPromise(const CancelableRef& cancelable, const PromiseRef<T,E>& promise, const SchedulerRef& sched) noexcept {
+    auto deferred = Task<None,None>::deferAction([cancelable](auto sched) {
+            return Deferred<None,None>::forCancelable(cancelable, sched);
+        })
+        .template flatMapBoth<T,E>(
+            [promise](auto) { return Task<T,E>::forPromise(promise); },
+            [promise](auto) { return Task<T,E>::forPromise(promise); }
+        )
+        .run(sched);
+
+    deferred->onCancel([promise]() {
+        promise->cancel();
+    });
+
     promise->onCancel([cancelable]() {
         cancelable->cancel();
     });
 
-    Task<None,None> cancelableDeferredTask = Task<None,None>::deferAction([cancelable](auto sched) {
-        return Deferred<None,None>::forCancelable(cancelable, sched);
-    });
-
-    Task<T,E> resultTask = Task<T,E>::forPromise(promise);
-
-    Task<T,E> composedTask = cancelableDeferredTask
-        .template flatMapBoth<T,E>(
-            [resultTask](auto) { return resultTask; },
-            [resultTask](auto) { return resultTask; }
-        );
-
-    return composedTask;
+    return deferred;
 }
 
 template <class T, class E>